Move game graphics implementation into graphics package
