+++ game/graphics/internal/mesh.go
@@ -1,0 +1,9 @@
+package internal
+
+const (
+	CoordAttributeIndex    = 0
+	NormalAttributeIndex   = 1
+	TangentAttributeIndex  = 2
+	TexCoordAttributeIndex = 3
+	ColorAttributeIndex    = 4
+)
