+++ game/graphics/sky.go
@@ -2,21 +2,35 @@ package graphics
 
 import "github.com/mokiat/gomath/sprec"
 
-// Sky represents the scene background.
-type Sky interface {
+func newSky() *Sky {
+	return &Sky{}
+}
 
-	// BackgroundColor returns the color of the background.
-	BackgroundColor() sprec.Vec3
+// Sky represents the Scene's background.
+type Sky struct {
+	backgroundColor sprec.Vec3
+	skyboxTexture   *CubeTexture
+}
 
-	// SetBackgroundColor changes the color of the background.
-	SetBackgroundColor(color sprec.Vec3)
+// BackgroundColor returns the color of the background.
+func (s *Sky) BackgroundColor() sprec.Vec3 {
+	return s.backgroundColor
+}
 
-	// Skybox returns the cube texture to be used as the background.
-	// If one has not been set, this method returns nil.
-	Skybox() CubeTexture
+// SetBackgroundColor changes the color of the background.
+func (s *Sky) SetBackgroundColor(color sprec.Vec3) {
+	s.backgroundColor = color
+}
 
-	// SetSkybox sets a cube texture to be used as the background.
-	// If nil is specified, then a texture will not be used and instead
-	// the background color will be drawn instead.
-	SetSkybox(skybox CubeTexture)
+// 	// Skybox returns the cube texture to be used as the background.
+// 	// If one has not been set, this method returns nil.
+func (s *Sky) Skybox() *CubeTexture {
+	return s.skyboxTexture
+}
+
+// SetSkybox sets a cube texture to be used as the background.
+// If nil is specified, then a texture will not be used and instead
+// the background color will be drawn instead.
+func (s *Sky) SetSkybox(skybox *CubeTexture) {
+	s.skyboxTexture = skybox
 }
